feat/GLGDI: add 2D texture support to the OpenGL backend and fix minor issues with OpenGL state changes

+++ Source/Skyrocket/Graphics/Renderer/OpenGL/GLResource.hpp
@@ -58,7 +58,7 @@ struct GLUniformInfo {
 struct GLProgram {
     GLuint id;
 
-    size_t num_uniforms;
+    GLint num_uniforms;
     GLUniformInfo* uniforms;
 
     GLProgram()
